Highlight "expired" members
This isn’t actually correct — it highlights members whose last
agreement date is before today. This is one idea for doing this on the
client side (maybe it would just be easier to calculate whether they’re
expired or not from the server).
Obviously better styling is also needed here!